Attach retry instructions to release automation build as a markdown doc (#51)
* Show clear retry instructions in markdown

Upload a doc describing how to retry to the AzDO build as a "summary" of the build. This makes AzDO display the subset-of-markdown document its own tab on the build summary screen, which is a lot easier to get to than digging into the logs of a specific build step to find the instructions. This is also an opportunity to filter out unnecessary instructions.

Update release docs to match.

Include a direct link to the page showing the retry doc in status update release issue comments.

Fix an inconsistent "." in the Report Failure step vs. the Report Success step.

* Use LogCmd prefix, point to new doc in comments

Rename AzDO logging command funcs to use LogCmd prefix and follow the AzDO command naming more closely. Update comments: improve descriptions and link to the new logging command doc page rather than the old one.